requirements.txt
# ============================================================================
# SARA AI CALLING BOT - Requirements
# ============================================================================
# Complete dependency list with versions tested and working
# Auto-installer available: dependencies are checked on startup
# ============================================================================

# Core Framework & Environment
# ----------------------------------------------------------------------------
python-dotenv==1.0.1          # Environment variable management
flask==3.1.3                  # Web framework for bot server
requests==2.32.3              # HTTP library for API calls
psutil==5.9.8                 # System and process utilities

# AI & Language Models
# ----------------------------------------------------------------------------
openai==1.51.2                # OpenAI API (GPT-4o-mini, TTS)
google-generativeai==0.8.3    # Google Gemini (alternative AI)

# Speech Recognition (STT)
# ----------------------------------------------------------------------------
faster-whisper==1.0.3         # High-performance speech-to-text

# Text-to-Speech (TTS)
# ----------------------------------------------------------------------------
gtts==2.5.3                   # Google TTS (fallback)

# Phone Call Integration
# ----------------------------------------------------------------------------
twilio==9.2.3                 # Twilio API for phone calls

# Audio Processing
# ----------------------------------------------------------------------------
numpy==2.1.1                  # Numerical arrays for audio
sounddevice==0.4.7            # Audio I/O
soundfile==0.12.1             # Audio file reading/writing
pygame==2.5.2                 # Audio playback
pyaudio==0.2.14               # Audio streaming
wave==0.0.2                   # WAV file handling

# Real-time Voice Processing & Interruption Detection
# ----------------------------------------------------------------------------
webrtcvad==2.0.10             # Voice activity detection
librosa==0.10.1               # Audio analysis
websockets==12.0              # WebSocket support

# Text Processing & Language Support
# ----------------------------------------------------------------------------
indic-transliteration==2.3.75 # Hindi/Hinglish transliteration
regex==2025.9.18              # Advanced regex for language detection

# HTTP & Networking
# ----------------------------------------------------------------------------
httpx==0.27.2                 # Modern HTTP client
httpcore==1.0.5               # HTTP core library

# WhatsApp Integration (Optional - for payment links & messaging)
# ----------------------------------------------------------------------------
fastapi>=0.104.0              # FastAPI for WhatsApp microservice
uvicorn>=0.24.0               # ASGI server for FastAPI
pydantic>=2.5.0               # Data validation for API models
razorpay>=1.4.1               # Razorpay payment links integration
motor>=3.3.0                  # Async MongoDB driver (optional, for persistence)
pymongo>=4.6.0                # MongoDB driver

# Optional (not required but recommended)
# ----------------------------------------------------------------------------
asyncio-mqtt==0.16.1          # MQTT for messaging (optional)
pjsua2==2.13.1                # SIP client (optional)
threading-timer==0.1.0        # Timer utilities (optional)

# ============================================================================
# Installation Instructions
# ============================================================================
# 
# AUTO-INSTALL (Recommended):
#   python main.py
#   Dependencies are automatically checked and installed
#
# MANUAL INSTALL:
#   pip install -r requirements.txt
#
# INDIVIDUAL INSTALL:
#   pip install <package-name>
#
# ============================================================================
# Note: Some packages may require system dependencies
# ============================================================================
# 
# PyAudio (Windows): May require Microsoft C++ Build Tools
#   Download: https://visualstudio.microsoft.com/visual-cpp-build-tools/
#
# PyAudio (macOS): Install portaudio first
#   brew install portaudio
#   pip install pyaudio
#
# PyAudio (Linux): Install portaudio19-dev
#   sudo apt-get install portaudio19-dev
#   pip install pyaudio
#
# ============================================================================

